Added the script part of it
I added the script part of it, the .cs file because that's actually where you can modify it and change the app. Although the process of making the app isn't the same so I gotta do something about that and also you can't see that in the main download due to it being the script and the players won't need that, or do they? I want to make sure that the modders get what they want without making it more confusing to the users.